[move]コード修正
