--- features/components/test.css
@@ -1,11 +1,0 @@
-/* CountDisplay.module.css */
-
-.count-display {
-    font-size: 24px;
-    font-weight: bold;
-    color: #333;
-    padding: 10px;
-    border: 2px solid #ccc;
-    border-radius: 4px;
-    margin-top: 20px;
-  }
